move NewTokenHandler callaback in itw own file
