+++ core/cat/looking_glass/agent_manager.py
@@ -6,22 +6,13 @@ from typing import List, Dict
 from langchain.docstore.document import Document
 from langchain.prompts import PromptTemplate
 from langchain.chains import LLMChain
-from langchain.callbacks.base import BaseCallbackHandler
 from langchain.agents import AgentExecutor, LLMSingleActionAgent
 
 from cat.looking_glass import prompts
+from cat.looking_glass.callbacks import NewTokenHandler
 from cat.looking_glass.output_parser import ToolOutputParser
 from cat.utils import verbal_timedelta
 from cat.log import log
-
-
-class NewTokenHandler(BaseCallbackHandler):
-
-    def __init__(self, cat):
-        self.cat = cat
-    def on_llm_new_token(self, token: str, **kwargs) -> None:
-        self.cat.send_ws_message(token, "chat_token")
-
 
 
 class AgentManager:
@@ -54,7 +45,11 @@ class AgentManager:
         )
 
         # main chain
-        agent_chain = LLMChain(prompt=prompt, llm=self.cat._llm, verbose=True)
+        agent_chain = LLMChain(
+            prompt=prompt,
+            llm=self.cat._llm,
+            verbose=True
+        )
 
         # init agent
         agent = LLMSingleActionAgent(
